pyproject.toml
[tool.poetry]
name = "cognite-data-model"
version = "0.1.0"
description = ""
authors = ["Angela Dang Atkinson"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^2.0.1"
arrow = "^1.2.3"
cognite-sdk = "^6.1.9"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
